requirements.txt
boto3
datasets
huggingface_hub
python-dotenv
pandas

